Clean up selection direction handling
This fixes the issue noted in
https://github.com/whatwg/html/pull/2437#issuecomment-286628558 where
the spec sometimes assumed it was OK to set the selection direction to
none, even though not all platforms support this. It introduces a new
concept, "set the selection direction", which will convert "none" to
"forward" as appropriate.

This commit also makes a couple editorial tweaks:

* Replaces the italicized selection direction states with simple
  strings, which allows us to avoid a switch statement to translate
  between them.
* Splits out getter and setters in <input> and <textarea>'s value IDL
  attribute definitions, and gives the setters a proper numbered list.