Standardize special behavior of colspan/rowspan
Fixes #1198. Browsers all behave differently here, the spec matches
none. The basic behavior selected was that of Chrome, except for very
high values where we don't spec its overflow behavior. The colspan
limit of 1000 matches Firefox/Edge (Firefox is not willing to raise its
limit), instead of higher limits of 8190 in Chrome or more in WebKit.
The rowspan limit of 65534 matches Firefox and Chrome, as opposed to
65535 in Edge or 8190 in WebKit.

For setting, I followed Firefox because it's the simplest. Chrome
converts negative numbers to 0, and Edge throws in some cases and clamps
in others.

Tests: https://github.com/whatwg/html/pull/1993